refactor: separated lib into separate modules
