+++ src/main.rs
@@ -1,4 +1,4 @@
-use aaprop::interface::create_router;
+use aaprop::create_router;
 
 #[shuttle_runtime::main]
 async fn axum() -> shuttle_axum::ShuttleAxum {
